% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noppa.R
\name{noppa}
\alias{noppa}
\title{Noppaheitto}
\usage{
noppa(lkm = 1, sivuja = 20, plot = FALSE, takaisinpalautus = TRUE)
}
\arguments{
\item{lkm}{Heittojen lukumäärä (oletuksena 1).}

\item{sivuja}{Nopan sivujen lukumäärä (oletuksena 6).}

\item{plot}{Tehdäänkö kuvaaja vai ei (oletuksena FALSE).}

\item{takaisinpalautus}{Onko nopanheitot takaisinpalautuvia, eli toisistaan riippumattomia (oletuksena TRUE).}
}
\value{
Tulostaa nopanheitot ja niiden summan, sekä summan pistetodennäköisyysfunktion.
}
\description{
Heittää k kertaa n sivuista noppaa, Heitot, niiden summa ja todennäköisyys tulostetaan.
}
\examples{
# Heitä 2 kertaa perinteistä 6 sivuista noppaa.
noppa(2,6)

}
