% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noppa.R
\name{noppa}
\alias{noppa}
\title{Noppaheitto}
\usage{
noppa(sivuja = 6, lkm = 1, takaisinpalautus = TRUE)
}
\arguments{
\item{sivuja}{Nopan sivujen lukumäärä (oletuksena 6).}

\item{lkm}{Heittojen lukumäärä (oletuksena 1).}

\item{takaisinpalautus}{Onko nopanheitot takaisinpalautuvia, eli toisistaan riippumattomia (oletuksena TRUE).}
}
\value{
Tulostaa nopanheitot ja niiden summan.
}
\description{
Simuloi nopanheiton.
}
\examples{
# Heitä noppaa 5 kertaa
noppa(6, 5)

}
